+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Changelog
 
+## [0.30.0] - 2024-10-31
+
+### Changed
+
+- Update Bitmovin's native Android SDK version to `3.90.0`
+- Update Bitmovin's native iOS SDK version to `3.77.0`
+
+### Added
+
+- `WebUiConfig.variant` to set the UI variant that should be used by the Bitmovin Web UI
+
+### Fixed
+
+- Spatial navigation in the Web UI does not work properly
+
+## [0.29.0] - 2024-09-09
+
+### Added
+
+- Android: `AdaptationConfig.initialBandwidthEstimateOverride` to override the selection of the optimal media tracks before actual bandwidth data is available
+
+### Changed
+
+- Update Bitmovin's native Android SDK version to `3.82.0`
+- Update Bitmovin's native iOS SDK version to `3.71.0`
+- Update IMA SDK dependency on Android to `3.33.0`
+
+### Changed
+
+## [0.28.0] - 2024-08-07
+
+### Added
+
+- Android: `AdItem.preloadOffset` to control when ad manifest gets loaded
+
+### Changed
+
+- Update Bitmovin's native Android SDK version to `3.78.0`
+- Update Bitmovin's native iOS SDK version to `3.67.0`
+
 ## [0.27.1] - 2024-07-05
 
 ### Fixed